Refactor Billing modal
